Add support for Dock battery
The TF201 is not able to show the docks battery (battery in detachable keyboard).
Jorge Ruesga wrote a patch for this (see [1]), but it never made its way into the TF201 configuration. According to him the attached patch should make it work (untested). The change makes it identical to the TF300t and TF700 configurations ([2] and ]3]).

[1] http://review.cyanogenmod.org/#/c/31040/
[2] https://github.com/CyanogenMod/android_device_asus_tf300t/blob/cm-10.2/overlay/frameworks/base/core/res/res/values/config.xml
[3] https://github.com/CyanogenMod/android_device_asus_tf700t/blob/cm-10.2/overlay/frameworks/base/core/res/res/values/config.xml